Item updater fix

+++ src/main/resources/db/migration/V1428437260__snapshot_bid.sql
@@ -1,8 +1,8 @@
 CREATE TABLE snapshot_bid (
-  bidId      BIGINT,
+  bidId      INTEGER,
   snapshotId INTEGER,
   CONSTRAINT snapshot_bid_bid_fk FOREIGN KEY (bidId) REFERENCES bids (id),
   CONSTRAINT snapshot_bid_snapshot_fk FOREIGN KEY (snapshotId) REFERENCES snapshots (id)
 );
 
-CREATE UNIQUE INDEX snapshot_bid_uniq_index ON snapshot_bid(bidId, snapshotId);
+CREATE UNIQUE INDEX snapshot_bid_uniq_index ON snapshot_bid (bidId, snapshotId);
